feat: Update routing to redirect to the home page; refactor components for improved structure and functionality - Changed redirect paths in updatePassword, layout, and various auth pages to root ("/") instead of "/home". - Added new dashboard components: HomeSwiper, MomentsSwiper, OffersContainer, and Product components. - Refactored existing components to improve organization and code readability. - Implemented dynamic menu items for better navigation. - Updated sitemap generation to ensure correct URL encoding.
