+++ src/app/(adminDashboard)/layout.tsx
@@ -10,7 +10,7 @@ export default async function Layout({
 }) {
   const { data: user } = await getProfile();
   if (!user?.is_admin) {
-    redirect("/home");
+    redirect("/");
   }
   return (
     <div className="flex h-full min-h-screen overflow-x-hidden">
