+++ src/app/(auth)/forget_password/page.tsx
@@ -5,7 +5,7 @@ import ForgotPasswordForm from "./ui/ForgotPasswordForm";
 
 export default async function Page() {
   const { session } = await getSession();
-  if (session) redirect("/home");
+  if (session) redirect("/");
   return (
     <div className="bg-gray-gradient relative box-border flex h-screen w-screen flex-col items-center justify-center overflow-hidden rounded-sm pb-[3px] text-left shadow-[0_8px_14px_3px_#000]">
       <ForgotPasswordForm />
