+++ src/app/(auth)/login/page.tsx
@@ -5,7 +5,7 @@ import React from "react";
 
 export default async function Page() {
   const { session } = await getSession();
-  if (session) redirect("/home");
+  if (session) redirect("/");
   return (
     <div className="flex min-h-screen items-center justify-center">
       <div className="z-10 w-full">
