+++ src/app/(dashboard)/products/[slug]/ui/BreadCrumbs.tsx
@@ -14,7 +14,7 @@ export default function BreadCrumbs() {
     <div  dir={translation?.default_language ==="ar" ? "rtl" : "ltr"} className="flex w-full items-center justify-center bg-gray-100">
       <div className="flex h-[3rem] w-full max-w-[75rem] flex-row items-center justify-start gap-3 px-6 text-sm capitalize">
         <Link
-          href={"/home"}
+          href={"/"}
           className="cursor-pointer leading-6 transition-all ease-linear hover:font-medium hover:text-slate-500 hover:underline"
         >
           {translation?.lang["home"]}
