+++ src/app/(dashboard)/ui/home/ui/Offers/Offer.tsx
