+++ src/app/(dashboard)/ui/home/ui/Offers/Offers.tsx
