+++ src/app/(dashboard)/ui/home/ui/ProductsSection/ProductsSection.tsx
