--- src/app/page.tsx
@@ -1,4 +1,0 @@
-import { redirect } from "next/navigation";
-export default async function Page() {
-  redirect("/home");
-}
